Devided interfaces 'cuz of MyBaits
