+++ src/main/java/com/solvd/persistence/interfaces/AirlineRepository.java
@@ -1,0 +1,7 @@
+package com.solvd.persistence.interfaces;
+
+import com.solvd.model.Airline;
+import com.solvd.persistence.utility.Repositorable;
+
+public interface AirlineRepository extends Repositorable<Airline> {
+}
