+++ src/main/java/com/solvd/persistence/interfaces/AirportRepository.java
@@ -1,0 +1,7 @@
+package com.solvd.persistence.interfaces;
+
+import com.solvd.model.Airport;
+import com.solvd.persistence.utility.Repositorable;
+
+public interface AirportRepository extends Repositorable<Airport> {
+}
