+++ src/main/java/com/solvd/persistence/interfaces/FlightRepository.java
@@ -1,0 +1,7 @@
+package com.solvd.persistence.interfaces;
+
+import com.solvd.model.Flight;
+import com.solvd.persistence.utility.Repositorable;
+
+public interface FlightRepository extends Repositorable<Flight> {
+}
