+++ src/main/resources/mappers/AirlineMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.solvd.persistence.AirlineRepository">
+<mapper namespace="com.solvd.persistence.interfaces.AirlineRepository">
 
     <insert id="create" parameterType="com.solvd.model.Airline">
         INSERT INTO airline (id, name)
