+++ src/main/resources/mappers/AirportMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.solvd.persistence.AirportRepository">
+<mapper namespace="com.solvd.persistence.interfaces.AirportRepository">
 
     <insert id="create" parameterType="com.solvd.model.Airport">
         INSERT INTO airport (id, name, latitude, longitude)
