+++ src/main/resources/mappers/FlightMapper.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.solvd.persistence.FlightRepository">
+<mapper namespace="com.solvd.persistence.interfaces.FlightRepository">
 
     <insert id="create">
         INSERT INTO flight (id, name, start_id, destination_id, airline_id, price)
-        VALUES (#{id}, #{name}, #{startId}, #{destinationId}, #{airlineId}, #{price})
+        VALUES (#{id}, #{name}, #{start}, #{destination}, #{airline}, #{price})
     </insert>
 
     <select id="loadById" resultMap="FlightResultMap">
@@ -20,11 +20,8 @@
         <id property="id" column="id"/>
         <result property="name" column="name"/>
         <result property="price" column="price"/>
-        <association property="startId"
-                     resultMap="com.solvd.persistence.utility.Repositorable.AirportResultMap"/>
-        <association property="destinationId"
-                     resultMap="com.solvd.persistence.utility.Repositorable.AirportResultMap"/>
-        <association property="airlineId"
-                     resultMap="com.solvd.persistence.utility.Repositorable.AirlineResultMap"/>
+        <association property="start" resultMap="com.solvd.persistence.interfaces.AirportRepository.AirportResultMap"/>
+        <association property="destination" resultMap="com.solvd.persistence.interfaces.AirportRepository.AirportResultMap"/>
+        <association property="airline" resultMap="com.solvd.persistence.interfaces.AirlineRepository.AirlineResultMap"/>
     </resultMap>
 </mapper>
